# backend/app-boot/src/test/resources/openapi-snapshot.json
{"components":{"schemas":{"CiteBody":{"properties":{"claim":{"type":"string"},"transcriptKey":{"type":"string"}},"type":"object"},"ConsentBody":{"properties":{"state":{"type":"string"},"subjectRef":{"type":"string"}},"type":"object"},"ContextDto":{"properties":{"aiAssistanceDisclosureRef":{"type":"string"},"citationCriterion":{"additionalProperties":{"type":"string"},"type":"object"},"consentRefs":{"items":{"type":"string"},"type":"array"},"criteria":{"items":{"$ref":"#/components/schemas/CriterionDto"},"type":"array"},"generatorVersionRef":{"type":"string"},"locale":{"type":"string"},"redactionPolicyRef":{"type":"string"},"redactionRunRef":{"type":"string"},"retentionPolicyRef":{"type":"string"},"rubricVersionRef":{"type":"string"},"schemaDigest":{"type":"string"},"signatureRef":{"type":"string"},"timezone":{"type":"string"},"wormChainRefs":{"items":{"type":"string"},"type":"array"}},"type":"object"},"CriterionDto":{"properties":{"criterionId":{"type":"string"},"jobRelatednessRationaleRef":{"type":"string"}},"type":"object"},"DsarBody":{"properties":{"reasonCode":{"type":"string"},"subjectRef":{"type":"string"}},"type":"object"},"ErasureBody":{"properties":{"dsarKey":{"type":"string"},"scope":{"$ref":"#/components/schemas/ScopeDto"}},"type":"object"},"ExportBody":{"properties":{"caseKey":{"type":"string"},"citationKeys":{"items":{"type":"string"},"type":"array"},"context":{"$ref":"#/components/schemas/ContextDto"}},"type":"object"},"FinalizeBody":{"properties":{"caseKey":{"type":"string"},"decisionOutcomeRef":{"type":"string"}},"type":"object"},"OpenBody":{"properties":{"aiOutputVersionRef":{"type":"string"},"sourceEvidenceRefs":{"items":{"type":"string"},"type":"array"}},"type":"object"},"ScopeDto":{"properties":{"citationKeys":{"items":{"type":"string"},"type":"array"},"exportArtifactKeys":{"items":{"type":"string"},"type":"array"},"reviewCaseKeys":{"items":{"type":"string"},"type":"array"},"tombstoneTargetEvidenceIds":{"items":{"type":"string"},"type":"array"},"transcriptKeys":{"items":{"type":"string"},"type":"array"}},"type":"object"},"TranscribeBody":{"properties":{"sourceObjectKey":{"type":"string"}},"type":"object"},"TransitionBody":{"properties":{"action":{"type":"string"},"caseKey":{"type":"string"},"oversightRoleRef":{"type":"string"},"ref":{"type":"string"}},"type":"object"}}},"info":{"title":"OpenAPI definition","version":"v0"},"openapi":"3.1.0","paths":{"/api/v1/interviews/{interviewId}/citations":{"post":{"operationId":"citeClaim","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CiteBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["citation-api-controller"]}},"/api/v1/interviews/{interviewId}/dsar":{"post":{"operationId":"receiveDsar","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DsarBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["dsar-api-controller"]}},"/api/v1/interviews/{interviewId}/dsar/erasure":{"post":{"operationId":"executeErasure","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErasureBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["dsar-api-controller"]}},"/api/v1/interviews/{interviewId}/export":{"post":{"operationId":"exportPacket","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExportBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["export-api-controller"]}},"/api/v1/interviews/{interviewId}/recording-consent":{"put":{"operationId":"putRecordingConsent","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}},{"in":"header","name":"X-ATS-Idempotency-Key","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConsentBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["interview-api-controller"]}},"/api/v1/interviews/{interviewId}/recordings":{"post":{"operationId":"uploadRecording","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}},{"in":"header","name":"X-ATS-Filename","required":false,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["interview-api-controller"]}},"/api/v1/interviews/{interviewId}/review-case":{"get":{"operationId":"getCase","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}},{"in":"query","name":"case","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["review-api-controller"]}},"/api/v1/interviews/{interviewId}/review-case/finalize":{"post":{"operationId":"finalizeDecision","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FinalizeBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["review-api-controller"]}},"/api/v1/interviews/{interviewId}/review-case/transition":{"post":{"operationId":"transition","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransitionBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["review-api-controller"]}},"/api/v1/interviews/{interviewId}/review-cases":{"get":{"operationId":"listCases","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["review-api-controller"]},"post":{"operationId":"open","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OpenBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["review-api-controller"]}},"/api/v1/interviews/{interviewId}/transcribe":{"post":{"operationId":"transcribe","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TranscribeBody"}}},"required":true},"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["interview-api-controller"]}},"/api/v1/interviews/{interviewId}/transcript":{"get":{"operationId":"getTranscript","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}},{"in":"query","name":"key","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["interview-api-controller"]}},"/api/v1/interviews/{interviewId}/transcripts":{"get":{"operationId":"listTranscripts","parameters":[{"in":"path","name":"interviewId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"content":{"*/*":{"schema":{"type":"object"}}},"description":"OK"}},"tags":["interview-api-controller"]}},"/healthz":{"get":{"operationId":"healthz","responses":{"200":{"content":{"*/*":{"schema":{"additionalProperties":{"type":"string"},"type":"object"}}},"description":"OK"}},"tags":["health-controller"]}}}}